compilación de less en cliente

+++ c_modules/mediaserver/classes/controller/CacheUtilsController.php
@@ -114,14 +114,16 @@ class CacheUtilsController {
 
     global $CACHE_UTILS_LESS_TMPDIR;
 
-    if(!$dir)
-       $dir = $CACHE_UTILS_LESS_TMPDIR;      
+    if( $CACHE_UTILS_LESS_TMPDIR ) {
+      if(!$dir)
+         $dir = $CACHE_UTILS_LESS_TMPDIR;      
 
-    $files = array_diff(scandir($dir), array('.','..'));
-    foreach ($files as $file) {
-      (is_dir("$dir/$file")) ? self::removeLessTmpdir("$dir/$file") : unlink("$dir/$file");
+      $files = array_diff(scandir($dir), array('.','..'));
+      foreach ($files as $file) {
+        (is_dir("$dir/$file")) ? self::removeLessTmpdir("$dir/$file") : unlink("$dir/$file");
+      }
+      rmdir($dir);
     }
-    rmdir($dir);
   }
